Implement pytest as GitHub action
This adds pytest as workflow via GitHub actions as enabler for unit testing
the code base. Test cases for the existing code base will have to be developed
later on.

There is currently a caveat: The GitHub action is triggered by `push` and by
`pull_request` events, resulting in duplicated runs when creating a pull
request from a branch that has been pushed to before. This will result in
triggering the unit tests twice in some cases, which might result in two
comments being posted / emails being sent.

Currently not sure how to properly prevent this, without skipping the tests
in some cases.