Implement reporting of test coverage via pytest-cov
This implements reporting of test coverage for the luxtronik module. The
information itself is provided via `pytest-cov` and is then uploaded as
comment via the `MishaKav/pytest-coverage-comment` GitHub action in the
`pytest` workflow.

It will create some artifacts that are used to exchange information between
the different steps, e.g. `pytest.xml` and `pytest-coverage.txt`.